+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 ## Changed
 - Remove modifier from AndesBadge
 
+## Fixed
+- Label error textfield color.
+- Helper and counter superpose.
+- Error icon only showed with a helper text.
+
 # v2.1.0
 ## Added
 - Badge Component. (Credits)